examples/fishing/breaching/requirements.txt
hydra-core
torch
torchvision
numpy
matplotlib
scipy
lpips
python-socketio
aiohttp
boto3
datasets
transformers
opacus
gym
zstd
torch-optimizer
timm